requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.12.2
certifi==2022.9.14
h11==0.13.0
idna==3.4
mysql-connector-python==8.0.30
outcome==1.2.0
protobuf==3.20.1
PySocks==1.7.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
speedtest-cli==2.1.3
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
xvfbwrapper==0.2.9
PyVirtualDisplay~=3.0
requests~=2.28.2
glob2~=0.7
numpy~=1.24.2
pandas~=1.3.5
pycurl~=7.45.2
fake-useragent==1.4.0
stem==1.8.2
